# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlogSphere API",
    "description": "API Documentation for BlogSphere application",
    "version": "1.0.0"
  },
  "host": "blog-sphere-backend-ruby.vercel.app",
  "schemes": ["https"],
  "basePath": "/",
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter token in format: Bearer <token>"
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "status": { "type": "integer", "example": 200 },
        "data": { "type": "object" },
        "message": { "type": "string", "example": "Operation successful" }
      }
    },
...
  },
  "paths": {
...
    "/api/v1/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get a single post",
        "description": "Retrieves a post by its ID.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" }
              }
            }
          },
          "404": { "description": "Post not found" }
        }
      },
      "patch": {
        "tags": ["Posts"],
        "summary": "Update a post",
        "description": "Updates an existing post with optional image upload. Requires authentication.",
        "security": [{ "BearerAuth": [] }],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post to update",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the post"
                  },
                  "content": {
                    "type": "string",
                    "description": "Content of the post"
                  },
                  "catagory": {
                    "type": "string",
                    "description": "Category of the post"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Comma-separated tags"
                  },
                  "contentTable": {
                    "type": "string",
                    "description": "Content table for the post"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional post image (e.g., PNG, JPEG)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Unauthorized (not post owner)" },
          "404": { "description": "Post not found" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete a post",
        "description": "Deletes a post by its ID. Requires authentication.",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post to delete",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" }
              }
            }
          },
          "403": { "description": "Unauthorized (not post owner)" },
          "404": { "description": "Post not found" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}